% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{assign_infection_history}
\alias{assign_infection_history}
\title{Assign infection history list to covfefe project}
\usage{
assign_infection_history(proj, infection_history)
}
\arguments{
\item{proj}{current covfefe project}

\item{infection_history}{TODO}
}
\description{
Assign infection history list to covfefe project
}
\details{
TODO
}
\examples{
# TODO
}
