% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bind_infected_durations}
\alias{bind_infected_durations}
\title{Bind infected durations to project}
\usage{
bind_infected_durations(proj, infected_durations)
}
\arguments{
\item{proj}{the current project}

\item{infected_durations}{list of infection lengths}
}
\description{
Bind infected durations to project
}
\examples{
# TODO
}
