% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bind_lags}
\alias{bind_lags}
\title{Bind mosquito lag durations to project}
\usage{
bind_lags(proj, lags)
}
\arguments{
\item{proj}{the current project}

\item{lags}{lag durations}
}
\description{
Bind mosquito lag durations to project. Values of -1 indicate that this infected host should have genotypes generated de novo, rather than from another host in this simulation.
}
\examples{
# TODO
}
