% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bind_migrations}
\alias{bind_migrations}
\title{Bind migration array to project}
\usage{
bind_migrations(proj, migrations)
}
\arguments{
\item{proj}{the current project}

\item{migrations}{migration array}
}
\description{
Bind migration array to project
}
\examples{
# TODO
}
