% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{covfefe_file}
\alias{covfefe_file}
\title{Import file}
\usage{
covfefe_file(name)
}
\arguments{
\item{name}{name of file}
}
\description{
Import file from the inst/extdata folder of this package
}
\examples{
# TODO
}
