% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{covfefe_infected_durations}
\alias{covfefe_infected_durations}
\title{Import infected durations}
\usage{
covfefe_infected_durations(name)
}
\arguments{
\item{name}{name of file}
}
\description{
Import infected durations from the inst/extdata folder of this package
}
\examples{
# TODO
}
