% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{covfefe_palette}
\alias{covfefe_palette}
\title{Use \code{\link{covfefe_color_palette}} to construct palettes of desired length.}
\usage{
covfefe_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{n}{Number of colours desired. Unfortunately most palettes now only
have 4 or 5 colours. But hopefully we'll add more palettes soon. 
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}

\item{number}{ID of desired palette (1 to 3)}
}
\value{
A vector of colours.
}
\description{
Use \code{\link{covfefe_color_palette}} to construct palettes of desired length.

Covfefe coachella palette generator
}
\examples{
covfefe_color_palette(2, 8)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- covfefe_palette(1, number = 2, type = "continuous")
image(volcano, col = pal)

}
\keyword{colours}
