% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfefe_project.R
\name{covfefe_project}
\alias{covfefe_project}
\title{Create covfefe project}
\usage{
covfefe_project(oocysts = dpois(0:9, lambda = 1), hepatocytes = dpois(0:9,
  lambda = 1))
}
\arguments{
\item{oocysts}{TODO}

\item{hepatocytes}{TODO}
}
\description{
Create covfefe project
}
\details{
TODO
}
\examples{
# TODO
}
