% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfefe.R
\name{covfefy}
\alias{covfefy}
\title{Covfefy any word.}
\usage{
covfefy(str = "coverage")
}
\arguments{
\item{str}{Character string of word to covfefy. Defaults to "coverage".}
}
\description{
We can't all be the Cheeto in Chief, so we need this function to ensure
that we have the best words. This function takes a word, and performs the
 following algorithm thanks to
 \href{https://codegolf.stackexchange.com/questions/123685/covfefify-a-string}{
 this codegolf thread}:

1. Include all characters up through the first vowel in the word. (co)

2. Identifies the next consonant after the first vowel in the word. (v)

3. Changes that consonant to a phonetically similar consonant. (f)

4. Finds the next vowel after that consonant. (e)

5. Combines the two, and repeats. (fefe)

6. Puts it all together: covfefe
}
