% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfefe.R
\name{covfefySentence}
\alias{covfefySentence}
\title{Covfefe-fy any sentence.}
\usage{
covfefySentence(sent = paste0("Despite the constant negative",
  " press coverage we are going ", "to Make America Great Again"),
  endSentence = FALSE)
}
\arguments{
\item{sent}{Character string of sentence to covfefy. Defaults to
the famous tweet.}

\item{endSentence}{Boolean where TRUE forces use of punctuation.}
}
\description{
Calls \code{covfefy()} and uses rules to decide which, if any, word is
covfefefied.
}
