% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfefe.R
\name{covfefySpeech}
\alias{covfefySpeech}
\title{Covfefe-fy any speech.}
\usage{
covfefySpeech(text = system.file("extdata", "inauguration.txt", package =
  "covfefe"), out = "covfefe_inauguration.txt")
}
\arguments{
\item{text}{Location of input .txt file.}

\item{out}{Location of output .txt file.}
}
\description{
Calls \code{covfefySentences()} and uses rules to decide which,
if any, sentences are covfefefied.
}
