% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_demography}
\alias{define_demography}
\title{Define simulation demographic parameters}
\usage{
define_demography(project, demography = dgeom(1:100, prob = 1/20))
}
\arguments{
\item{project}{covfefe project.}

\item{demography}{vector specifying proportion of the population in each
one-year age group.}
}
\description{
Define simulation demographic parameters
}
