% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_demograpy}
\alias{define_demograpy}
\title{Define simulation demographic parameters}
\usage{
define_demograpy(project, life_table = NULL)
}
\arguments{
\item{project}{covfefe project.}

\item{life_table}{vector specifying probability of death in each one-year age
group. Final value must be 1 to ensure a closed population.}
}
\description{
Define deography used in simulation. The raw input is a life
  table giving the probability of death in each one-year age group. This in
  turn is used to derive the distribution of age of death, and the stable age
  distribution.
}
