% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_epi_parameters}
\alias{define_epi_parameters}
\title{Define simulation epidemiological parameters}
\usage{
define_epi_parameters(project, a = 0.3, p = 0.9, mu = -log(p),
  u = 10, v = 10, g = 10, r = 1/200, b = 1, c = 1,
  max_innoculations = 5)
}
\arguments{
\item{project}{covfefe project.}

\item{a}{human blood feeding rate. The proportion of mosquitoes that feed on
humans each day.}

\item{p}{mosquito probability of surviving one day.}

\item{mu}{mosquito instantaneous death rate. mu = -log(p) unless otherwise
specified.}

\item{u}{intrinsic incubation period. The number of days from infection to
blood-stage infection in a human host.}

\item{v}{extrinsic incubation period. The number of days from infection to
becoming infectious in a mosquito.}

\item{g}{lag time between human blood-stage infection and production of
gametocytes.}

\item{r}{daily recovery rate.}

\item{b}{probability a human becomes infected after being bitten by an
infected mosquito.}

\item{c}{probability a mosquito becomes infected after biting an infected
human.}

\item{max_innoculations}{maximum number of innoculations that an individual
can hold simultaneously.}
}
\description{
Define simulation epidemiological parameters
}
