% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_hepatocyte_distribution}
\alias{define_hepatocyte_distribution}
\title{Define hepatocyte count distribution}
\usage{
define_hepatocyte_distribution(proj, hepatocytes)
}
\arguments{
\item{proj}{current covfefe project}

\item{hepatocytes}{TODO}
}
\description{
Define hepatocyte count distribution
}
\details{
TODO
}
\examples{
# TODO
}
