% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_migration}
\alias{define_migration}
\title{Define simulation migration parameters}
\usage{
define_migration(project, migration_matrix = matrix(1))
}
\arguments{
\item{project}{covfefe project.}

\item{migration_matrix}{TODO.}
}
\description{
Define simulation migration parameters
}
