% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_oocyst_distribution}
\alias{define_oocyst_distribution}
\title{Define oocyst count distribution}
\usage{
define_oocyst_distribution(proj, oocysts)
}
\arguments{
\item{proj}{current covfefe project}

\item{oocysts}{TODO}
}
\description{
Define oocyst count distribution
}
\details{
TODO
}
\examples{
# TODO
}
