% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{delete_infection_history}
\alias{delete_infection_history}
\title{Delete infection history list from covfefe project}
\usage{
delete_infection_history(proj)
}
\arguments{
\item{proj}{current covfefe project}
}
\description{
Delete infection history list from covfefe project
}
\details{
TODO
}
\examples{
# TODO
}
