% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.covfefe_indiv}
\alias{plot.covfefe_indiv}
\title{Plot \code{covfefe_indiv} object}
\usage{
\method{plot}{covfefe_indiv}(x, y, ...)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{...}{TODO}
}
\description{
Default plot for \code{covfefe_indiv} object.
}
\details{
TODO
}
\examples{
# TODO
}
