% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_innoculations}
\alias{plot_innoculations}
\title{Plot daily number of innoculated infections}
\usage{
plot_innoculations(x, deme = 1, ymax = NULL)
}
\arguments{
\item{x}{TODO}

\item{deme}{TODO}

\item{ymax}{TODO}
}
\description{
Plot daily number of hosts with each possible number of innoculated infections.
}
\details{
TODO
}
\examples{
# TODO
}
