% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ross_macdonald.R
\name{ross_macdonald}
\alias{ross_macdonald}
\title{RM1_stochastic_sync}
\usage{
ross_macdonald(max_time = 100, a = 0.3, p = 0.9, mu = -log(p), u = 22,
  v = 10, r = 1/200, b = 1, c = 1, Ih_init = 10, Iv_init = 0,
  H = 100, M = 100)
}
\arguments{
\item{max_time}{run simulation for this many days}

\item{a}{human blood feeding rate. The proportion of mosquitoes that feed on humans each day.}

\item{p}{mosquito probability of surviving one day.}

\item{mu}{mosquito instantaneous death rate. mu = -log(p) unless otherwise specified.}

\item{u}{intrinsic incubation period. The number of days from infection to becoming infectious in a human host.}

\item{v}{extrinsic incubation period. The number of days from infection to becoming infectious in a mosquito .}

\item{r}{daily recovery rate.}

\item{b}{probability a human becomes infected after being bitten by an infected mosquito.}

\item{c}{probability a mosquito becomes infected after biting an infected human.}

\item{Ih_init}{initial number of infectious humans.}

\item{Iv_init}{initial number of infectious mosquitoes.}

\item{H}{human population size.}

\item{M}{mosquito population size (number of adult female mosquitoes).}

\item{Eh_init}{initial number of infected but not infectious humans.}

\item{Ev_init}{initial number of infected but not yet infectious mosquitoes.}
}
\description{
Draws from synchronous stochastic Ross-Macdonald model
}
\details{
Note that this model is stochastic and uses a daily time step, whereas the classic Ross-Macdonald system is deterministic and assumes continuous time. The stochastic nature of this model means we may see fade-out at low transmission, causing the average behaviour to differ from the deterministic prediction. The daily time step also means results may not agree exactly with classic Ross-Macdonald model.
}
