% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sample_hosts}
\alias{sample_hosts}
\title{Sample hosts from infection history}
\usage{
sample_hosts(proj, samp_mat, demes, recover_pop_counts = FALSE,
  max_infections = 5)
}
\arguments{
\item{proj}{current covfefe project}

\item{samp_mat}{TODO}

\item{demes}{TODO}

\item{recover_pop_counts}{TODO}

\item{max_infections}{TODO}
}
\description{
Sample hosts from infection history
}
\details{
TODO
}
\examples{
# TODO
}
