% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_genotypes}
\alias{sim_genotypes}
\title{Simulate genotypes from pruned infection tree}
\usage{
sim_genotypes(proj, loci, recom_rate)
}
\arguments{
\item{proj}{current covfefe project}

\item{loci}{TODO}

\item{recom_rate}{TODO}
}
\description{
Simulate genotypes from pruned infection tree
}
\details{
TODO
}
\examples{
# TODO
}
