% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sim_plot_EIR_innoculations}
\alias{sim_plot_EIR_innoculations}
\title{Simulate and plot EIR vs. innoculations}
\usage{
sim_plot_EIR_innoculations(max_time = 365, a = 0.3, p = 0.9,
  mu = -log(p), u = 22, v = 10, g = 10, r = 1/20, b = 1,
  c = 1, H = 100, M = c(50, 100, 200, 500, 1000, 2000, 5000, 10000),
  max_innoculations = 5, reps = 10)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{...}{TODO}
}
\description{
Simulate and plot EIR vs. innoculations
}
\details{
TODO
}
\examples{
# TODO
}
