% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfefe_project.R
\name{covfefe_project}
\alias{covfefe_project}
\title{Create covfefe project}
\usage{
covfefe_project(recom_rate = 0.01, bloodstage_skew = 1,
  biting = dpois(0:9, lambda = 1), oocysts = dpois(0:9, lambda = 1),
  hepatocytes = dpois(0:9, lambda = 1))
}
\arguments{
\item{recom_rate}{TODO}

\item{bloodstage_skew}{TODO}

\item{biting}{TODO}

\item{oocysts}{TODO}

\item{hepatocytes}{TODO}
}
\description{
Create covfefe project
}
\examples{
# TODO
}
