% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_genotypes}
\alias{sim_genotypes}
\title{Simulate genotypes}
\usage{
sim_genotypes(line_list, samp_mat, demes)
}
\arguments{
\item{line_list}{TODO}

\item{samp_mat}{TODO}

\item{demes}{TODO}
}
\description{
Prune infection tree and simulate genotypes
}
\examples{
# TODO
}
