% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_genotypes}
\alias{sim_genotypes}
\title{Simulate genotypes}
\usage{
sim_genotypes(line_list, samp_mat, demes, loci, recom_rate)
}
\arguments{
\item{line_list}{TODO}

\item{samp_mat}{TODO}

\item{demes}{TODO}

\item{loci}{TODO}

\item{recom_rate}{TODO}
}
\description{
Prune infection tree and simulate genotypes
}
\details{
TODO
}
\examples{
# TODO
}
