% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfefe_indiv.R
\name{sim_indiv}
\alias{sim_indiv}
\title{Simulate from simple individual-based model}
\usage{
sim_indiv(max_time = 365, a = 0.3, p = 0.9, mu = -log(p), u = 22,
  v = 10, g = 10, r = 1/20, b = 1, c = 1, Ih_init = 10, H = 1000,
  M = 1000, max_infections = 5, migration_matrix = matrix(1),
  H_auto = FALSE)
}
\arguments{
\item{max_time}{run simulation for this many days}

\item{a}{human blood feeding rate. The proportion of mosquitoes that feed on humans each day.}

\item{p}{mosquito probability of surviving one day.}

\item{mu}{mosquito instantaneous death rate. mu = -log(p) unless otherwise specified.}

\item{u}{intrinsic incubation period. The number of days from infection to becoming infectious in a human host.}

\item{v}{extrinsic incubation period. The number of days from infection to becoming infectious in a mosquito.}

\item{g}{lag time between human bloodstage infection and production of gametocytes.}

\item{r}{daily recovery rate.}

\item{b}{probability a human becomes infected after being bitten by an infected mosquito.}

\item{c}{probability a mosquito becomes infected after biting an infected human.}

\item{Ih_init}{initial number of infectious humans in each deme.}

\item{H}{human population size in each deme.}

\item{M}{mosquito population size (number of adult female mosquitoes) in each deme.}

\item{max_infections}{maximum number of infections that an individual can hold simultaneously.}

\item{migration_matrix}{matrix giving the probability of migrating from current deme (rows) to new deme (columns).}

\item{H_auto}{if TRUE then human population sizes are chosen automatically based on the migration matrix. In this case the total number of humans is equal to sum(H).}
}
\description{
Simulate from simple individual-based model
}
\details{
TODO
}
\examples{
# TODO
}
